Se agrego registro , token, login

+++ Controllers/CommentController.cs
@@ -22,60 +22,80 @@ namespace api.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             var comments = await _commentRepo.GetAllAsync();
-            var commentDto = comments.Select(s=> s.ToCommentDto());
-           return Ok(commentDto); 
+            var commentDto = comments.Select(s => s.ToCommentDto());
+            return Ok(commentDto);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             var comments = await _commentRepo.GetByIdAsync(id);
-            if(comments == null)
+            if (comments == null)
             {
                 return NotFound();
             }
             return Ok(comments.ToCommentDto());
         }
 
-        [HttpPost("{stockId}")]
+        [HttpPost("{stockId:int}")]
         public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentDto commentDto)
         {
-            if(!await _stockRepo.StockExists(stockId))
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            if (!await _stockRepo.StockExists(stockId))
             {
                 return BadRequest("Stock does not exist");
             }
 
             var commentModel = commentDto.ToCommentFromCreate(stockId);
             await _commentRepo.CreateAsync(commentModel);
-            return CreatedAtAction(nameof(GetById), new {id = commentModel},commentModel.ToCommentDto());
+            return CreatedAtAction(nameof(GetById), new { id = commentModel }, commentModel.ToCommentDto());
         }
 
-                [HttpPut]
-                [Route("{id}")]
+        [HttpPut]
+        [Route("{id:int}")]
 
-                public async  Task<IActionResult>Update([FromRoute] int id,UpdateCommentRequestDto updateDto)
-                {
-                    var comment = await _commentRepo.UpdateAsync(id, updateDto.ToCommentFromUpdate());
-                    if(comment == null)
-                    {
-                        return NotFound();
-                    }
-                    return Ok(comment.ToCommentDto());
+        public async Task<IActionResult> Update([FromRoute] int id, UpdateCommentRequestDto updateDto)
+        {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            var comment = await _commentRepo.UpdateAsync(id, updateDto.ToCommentFromUpdate());
+            if (comment == null)
+            {
+                return NotFound();
+            }
+            return Ok(comment.ToCommentDto());
 
-                } 
+        }
 
-                [HttpDelete]
-                [Route("{id}")]
-                public async Task<IActionResult> Delete([FromRoute]int id)
-                {
-                    var commentModel = await _commentRepo.DeleteAsync(id);
-                    if(commentModel == null)
-                    {
-                        return NotFound("Comment dos not exist");
-                    }
-                    return Ok(commentModel);
-                }
+        [HttpDelete]
+        [Route("{id:int}")]
+        public async Task<IActionResult> Delete([FromRoute] int id)
+        {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            var commentModel = await _commentRepo.DeleteAsync(id);
+            if (commentModel == null)
+            {
+                return NotFound("Comment dos not exist");
+            }
+            return Ok(commentModel);
+        }
